Can now add to playlist from the currently playing song details panel
Also changed the appearance of the panel to fit the extra button. It is 30 pixels taller, the shuffle and loop buttons are now in a row with the new button instead of next to the seekbar, and the shuffle button goes grey instead of changing icon when it is disabled.